Split reconcile.go into several files
